Added url import functionality
